+++ src/main/java/com/b2b/dbservice/dao/SignUpDAO.java
@@ -4,7 +4,7 @@ import com.b2b.dbservice.model.DD_B2B_CUSTOMER;
 
 public interface SignUpDAO {
 
-	DD_B2B_CUSTOMER getUserbyId(Integer id);
+	DD_B2B_CUSTOMER getUserbyId(Long id);
 
 	DD_B2B_CUSTOMER signUp(DD_B2B_CUSTOMER user);
